backend/requirements.txt
# ── Core framework ────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9      # form / cookie parsing
python-dotenv>=1.0.0

# ── Pydantic / Settings ───────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0              # PostgreSQL async driver
aiosqlite>=0.20.0            # SQLite async driver (dev/test)
alembic>=1.13.0              # database migrations

# ── Auth & Encryption ─────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0   # JWT RS256
cryptography>=43.0.0               # AES-256-GCM for Kite token at rest

# ── Zerodha Kite Connect SDK ──────────────────────────────────────────────────
kiteconnect>=4.2.0

# ── Technical Analysis / KPI Engine ──────────────────────────────────────────
pandas>=2.2.0
pandas-ta>=0.3.14b0
numpy>=1.26.0

# ── Fundamental data scraping (NSE India) ────────────────────────────────────
httpx>=0.27.0                # async HTTP client
tenacity>=8.3.0              # retry logic for NSE scraping

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler>=3.10.0
pytz>=2024.1

# ── Rate limiting ─────────────────────────────────────────────────────────────
slowapi>=0.1.9

# ── XIRR computation ─────────────────────────────────────────────────────────
pyxirr>=0.10.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dateutil>=2.9.0
